+++ src/features/chart/index.js
@@ -1,3 +1,4 @@
 export { Chart } from "./components/Chart";
 export { updateChartAction, updateChartDataAction } from "./redux/actions";
 export { chartReducer } from "./redux/reducer";
+export { chartMiddleware } from "./redux/middleware";
