+++ src/features/search/index.js
@@ -1,1 +1,2 @@
 export { Search } from "./components/Search";
+export { searchMiddleware } from "./redux/middleware";
