+++ src/features/search/redux/middleware.js
@@ -1,0 +1,8 @@
+import { UPDATE_SEARCH_QUERY } from "../../../store/constants";
+
+export const searchMiddleware = socketService => store => next => action => {
+  if (action.type === UPDATE_SEARCH_QUERY) {
+    socketService.get().emit("searchQuery", action.payload);
+  }
+  return next(action);
+};
