+++ src/features/stock-ticker/index.js
@@ -1,3 +1,4 @@
 export { StockTicker } from "./components/StockTicker";
 export { updateStockTickerAction } from "./redux/actions";
 export { stockTickerReducer } from "./redux/reducer";
+export { stockMiddleware } from "./redux/middleware";
